requirements.txt
requests
beautifulsoup4
lxml
dnspython
googlesearch-python
scrapy
pandas
python-dotenv
loguru
shodan
dnspython
requests